requirements.txt
openai==0.28.0
torch==1.13.0
numpy==1.26.3
pandas==2.1.4
requests==2.31.0 
requests-oauthlib==1.3.1
seaborn==0.13.2
tiktoken>=0.5.0
python-dotenv>=1.0.0  # Optional: for .env file support